requirements.txt
flask
PyJWT
flask-session
python-dotenv
psycopg2-binary
bcrypt